pyproject.toml
[project]
name = "fastapi-architecture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.135.3",
    "greenlet>=3.4.0",
    "pillow>=12.2.0",
    "psycopg[binary]>=3.3.3",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "sqlalchemy>=2.0.49",
]
